requirements.txt
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.4
asn1crypto==0.24.0
astroid==2.5.2
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bleach==4.1.0
bootstrap4==0.1.0
Brlapi==0.6.6
bs4==0.0.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==6.7
colorama==0.4.4
cryptography==2.1.4
cupshelpers==1.0
dataclasses==0.8
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==3.2
django-bootstrap4==3.0.0
django-cors-headers==3.7.0
django-decouple==2.1
djangorestframework==3.14.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
findspark==1.4.2
httplib2==0.11.3
humanfriendly==9.1
idna==2.10
importlib-metadata==2.1.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.8.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kafka-python==2.0.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
louis==3.5.0
lxml==4.6.3
macaroonbakery==1.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
oauth==1.0.1
olefile==0.45.1
openpyxl==3.0.7
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==2.4.0
ply==3.11
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.4
psycopg2==2.8.6
ptyprocess==0.7.0
py4j==0.10.9.3
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.14.1
pycups==1.9.73
Pygments==2.10.0
PyGObject==3.26.1
pylint==2.7.4
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.1.2
pyparsing==3.0.6
pyreadline==2.1
pyRFC3339==1.0
pyrsistent==0.18.0
PySide2==5.15.2
pysmi==0.3.4
pysnmp==4.4.12
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
pytweening==1.0.4
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.2
reportlab==3.4.0
requests==2.25.1
requests-unixsocket==0.1.5
scapy==2.4.5
schedule==1.1.0
SecretStorage==2.3.1
Send2Trash==1.8.0
shiboken2==5.15.2
simplejson==3.13.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
system-service==0.3
systemd-python==234
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typed-ast==1.4.3
typing_extensions==4.1.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
usb-creator==0.3.3
vboxapi==1.0
virtualenv==20.7.2
virtualenvwrapper-win==1.2.6
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.12.1
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==1.3.7
zipp==3.6.0
zope.interface==4.3.2
